requirements.txt
accelerate==0.27.2
av==12.2.0
boto3==1.34.138
botocore==1.34.138
decord==0.6.0
diffusers==0.29.2
einops==0.4.1
imageio==2.21.1
imageio-ffmpeg==0.4.9
kornia==0.6.9
matplotlib==3.9.1
mediapipe==0.10.14
numpy==1.23.5
omegaconf
opencv_contrib_python==4.10.0.84
opencv_python==4.6.0.66
opencv_python_headless==4.10.0.84
Pillow==9.5.0
pycocotools==2.0.8
pyrender==0.1.45
PyYAML==6.0.1
rembg
Requests==2.32.3
safetensors==0.4.5
scikit_image==0.19.2
smplx==0.1.20
scipy==1.10.0
six==1.16.0
tensorboardX==2.6.2.2
termcolor==2.4.0
timm==0.9.7
tqdm==4.64.1
transformers==4.30.2
trimesh==4.4.1
xformers==0.0.22
yacs==0.1.8
kiui
git+https://github.com/NVlabs/nvdiffrast.git